refactor: only writing the logic for NE N-glycan core, using `convert_ne_to_dn()` if mode is "dn"
The rationale is that currently `n_glycan_core()` just build a "concrete" graph and uses `convert_glycan_mono_type()` to convert to other monosaccharide types, so it would be consistent if it also uses `convert_ne_to_dn()`. This function hasn't be created when `n_glycan_core()` was created, so it is a good time now to do the refactoring. Also, it would be good as well to make the "linkage" logic into the main function as well. That is, first building a graph with linkage, and if "linakge" is FALSE, delete the linkages. The internal function `n_glycan_core_ne()` now focusses on a simple job: creating a NE graph with "concrete" monosaccharides and full linkages.