feat: implement subsetting and slicing methods for `glyrepr_structure` class
The default "vctrs way" of doing this will copy the whole structure list attribute to the new object. This is not ideal as the new object might be much shorter and not contain all structures as before. This commit provides a `vec_restore.glyrepr_structure` method that handles the structure attributes properly.