refactor: refactor `structure_to_iupac`
The generation of IUPAC-condensed nomenclatures is split into two steps: 1. recursive generation of pseudo-IUPAC like V1E1[V2E2]V3E3; 2. generate real IUPAC by replace vertex and edge indices with monosaccharides and linkages.

This refactor is for paving the way for reordering the vertices and edges to be in line with IUPAC nomenclatures. The logic of the refactored first step can be used for this purpose.