feat: rename generic names of some monosaccharides
Some monosaccharides (e.g. Mur, Neu) have the same names for both generic and concrete mono-types. This can lead to ambiguity in deciding composition or structure mono-types. For example, we could not know whether "Mur(??-" is a generic or concrete one. This might be ok for explorative analysis, but may cause trouble for some functions that strictly rely on the mono-type of a glycan.

Now we add a "g" prefix to all the special monosaccharides. For example, concrete "Mur" now has a generic name "gMur".

This commit only update the monosaccharide table and some tests. Following commits will simplify some relevant functions and update the documentation.